requirements.txt
django-formtools
django>=1.11.0<=2.0
